+++ R/utils.R
@@ -28,12 +28,12 @@ default_gander_style <- function() {
 default_gander_dims <- c(5L, 100L)
 
 get_gander_style <- function() {
-  res <- getOption(".gander_style")
+  res <- getOption("gander.style", getOption(".gander_style"))
 
   if (!is.null(res)) {
     check_string(
       res,
-      arg = 'getOption(".gander_style")',
+      arg = 'getOption("gander.style")',
       call = call2("gander_addin")
     )
     return(res)